Update README.md
Correcting some faulty Markdown formatting and adding the "BACK TO INDEX OF REPOSITORIES" hyperlink plus the "A.G. Signature" for 2024, which includes a hyperlink to my alexgagnon.com personal website. I might want to change this to my new website for my art and research labs, called The Historiotheque, currently running at the historiotheque.ca website. A.G.